自动化代码提交

+++ MDCommonKit/Classes/UIColor+Ext.h
@@ -13,7 +13,7 @@
 + (UIColor *)md_colorWithHexString:(NSString *)hexString;
 + (UIColor *)md_colorWithHexString:(NSString *)hexString alpha:(CGFloat)alpha;
 + (UIColor *)md_colorWithIntRed:(NSInteger)r green:(NSInteger)g blue:(NSInteger)b;
-+ (UIColor *)md_colorWithIntRed:(NSInteger)r green:(NSInteger)g blue:(NSInteger)b alpha:(NSInteger)a;
++ (UIColor *)md_colorWithIntRed:(NSInteger)r green:(NSInteger)g blue:(NSInteger)b alpha:(CGFloat)a;
 
 @end
